requirements.txt
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
transformers==4.31.0
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
transformers==4.31.0
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests[socks]==2.31.0
PySocks==1.7.1
PyPDF2==3.0.1
pdfplumber==0.10.2
accelerate==0.21.0
einops==0.6.1
biopython==1.81
huggingface-hub==0.14.1
arxiv>=2.0.0
plotly>=5.18.0
dash>=2.14.2
selenium==4.15.2
webdriver_manager==4.0.1
unpywall==0.2.3
PySocks==1.7.1
PyPDF2==3.0.1
pdfplumber==0.10.2
accelerate==0.21.0
einops==0.6.1
biopython==1.81
huggingface-hub==0.14.1
arxiv>=2.0.0
plotly>=5.18.0
dash>=2.14.2
selenium==4.15.2
webdriver_manager==4.0.1
unpywall==0.2.3
